+++ packages/mcp/src/library.ts
@@ -1,19 +1,9 @@
-import './style.scss'
-
 // Classes
 export * from './classes/Mcp'
 export * from './classes/McpItemAbstract'
 export * from './classes/McpPrompt'
 export * from './classes/McpResource'
-export * from './classes/McpResourcesLoader'
 export * from './classes/McpTool'
-
-// Components
-
-// Composables
-
-// Functions
-export * from './functions/getMcpResources'
 
 // Types
 export * from './types/McpTypes'